plagiarism_checker/api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
faiss-cpu==1.7.4
openai==1.14.3
requests==2.31.0
python-dotenv==1.0.1
numpy==1.26.4
pydantic==1.10.13